# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tech/news": {
            "get": {
                "description": "Returns a list of news related to the tech genre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tech"
                ],
                "summary": "Get tech news",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.News"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tech/news/topic-of-the-season": {
            "get": {
                "description": "Returns the topic of the season for the tech genre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tech"
                ],
                "summary": "Get topic of the season for tech",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.NewsWithTopic"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tech/projects": {
            "get": {
                "description": "Returns a list of tech-related projects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tech"
                ],
                "summary": "Get projects related to tech",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TechProjects"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.News": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genre": {
                    "$ref": "#/definitions/models.NewsGenres"
                },
                "headline": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link-to-source": {
                    "type": "string"
                },
                "sentiment": {
                    "$ref": "#/definitions/models.NewsSentiments"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.NewsGenres": {
            "type": "string",
            "enum": [
                "tech",
                "gaming",
                "finance"
            ],
            "x-enum-varnames": [
                "NewsGenreTech",
                "NewsGenreGaming",
                "NewsGenreFinance"
            ]
        },
        "models.NewsSentiments": {
            "type": "string",
            "enum": [
                "undefined",
                "good",
                "indifferent",
                "bad"
            ],
            "x-enum-varnames": [
                "SentimentUndefined",
                "SentimentGood",
                "SentimentIndifferent",
                "SentimentBad"
            ]
        },
        "models.TechProjects": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link-to-git": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TimeframeTypes": {
            "type": "string",
            "enum": [
                "custom",
                "1d",
                "1w",
                "1m",
                "3m",
                "6m",
                "1y"
            ],
            "x-enum-varnames": [
                "TimeframeCustom",
                "Timeframe1D",
                "Timeframe1W",
                "Timeframe1M",
                "Timeframe3M",
                "Timeframe6M",
                "Timeframe1Y"
            ]
        },
        "models.TopicOfTheSeasons": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "custom_end": {
                    "type": "string"
                },
                "custom_start": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "genre": {
                    "$ref": "#/definitions/models.NewsGenres"
                },
                "id": {
                    "type": "integer"
                },
                "topic": {
                    "type": "string"
                },
                "topic_timestamp": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.TimeframeTypes"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "service.NewsWithTopic": {
            "type": "object",
            "properties": {
                "news_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.News"
                    }
                },
                "topic_of_the_season": {
                    "$ref": "#/definitions/models.TopicOfTheSeasons"
                }
            }
        }
    }
}